+++ src/dwarf/debug_link.rs
@@ -3,7 +3,7 @@
 //! From <https://sourceware.org/gdb/current/onlinedocs/gdb.html/Separate-Debug-Files.html>:
 //!
 //! A debug link is a special section of the executable file named
-//! .`gnu_debuglink`. The section must contain:
+//! `.gnu_debuglink`. The section must contain:
 //! - A filename, with any leading directory components removed, followed by a
 //!   zero byte,
 //! - zero to three bytes of padding, as needed to reach the next four-byte